requirements.txt
numpy~=1.17.3
chainer~=6.5.0
pychord~=0.4.0
Django~=2.2.6
django-allauth~=0.40.0
django-filter~=2.2.0
django-glrm~=1.1.3
djangorestframework~=3.10.3
jsonfield~=2.0.2
